keyboard_typing_game
Keyboard jump game is a speed typing game that helps in improving the typing speed of players

The object of Keyboard Jump Game Python Project is to build a keyboard jump game that helps players to increase their typing speed. We use pygame, random, and time modules in this project.

In this project, the player has to press the same keys to the letters displayed on the game screen. If the player made an error while typing then the game gets over.