Spinner
The games are cool and everyone loves them. Making games can be very different from other types of programming. In this article, I will walk you through how to create a game with the Python programming language.

Now let’s see how to create a game with Python. I am going to create a very simple game based on a fidget spinner. The logic of the game is that the turns will keep increasing as you press the space bar, and it will reduce its speed and stop at a point where you stop pressing the space bar

Hope now it clear why and  how you play it ...