Guess The Gibberish
What is Gibberish?
Gibberish, also called jibber-jabber or gobbledygook, is speech that is nonsense. It may include speech sounds that are not actual words, or language games and specialized jargon that seems nonsensical to outsiders.

Is gibberish a real language?

In addition to nonsense words, phrases and sentences, there is also a language called Gibberish. The language is similar to Pig Latin and is used by people who want to play games with a secret language. To speak the language, you break each word down into its syllables.

Hope you Understand Now see some Examples...
  see method_of_speech.jpg
for examples....


Thankyou..